move logging_context to tools
